pyproject.toml
[project]
name = "brownfield-cartographer"
version = "0.1.0"
description = "A codebase intelligence tool that analyzes repositories and maps their structure and data flow."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "langchain-core>=1.2.18",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.2",
    "networkx",
    "numpy>=2.2.6",
    "pydantic",
    "pyvis>=0.3.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "sqlglot",
    "streamlit>=1.55.0",
    "tree-sitter",
    "tree-sitter-python",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.6",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 50
